Monasca Events Publishing Spec
Monasca Events API was developed to store events
data in Elasticsearch.

There is still a need to collect and publish
Openstack Notifications as Monasca Events.

This commit includes a specification for creating a new Monasca
Events Listener service that would be independent of Ceilometer.
This is being proposed for the Stein cycle.

Monasca Ceilometer project currently publishes ceilometer samples
to Monasca API. An alternate specification proposes to extend Monasca
Ceilometer project and add a new events publisher which would publish
Openstack notifications (or events) to Monasca Events API.
This solution was proposed for Rocky but will not be implemented
as the Ceilometer Events feature has been deprecated on master.

Co-Authored-By: Joseph Davis <joseph.davis@suse.com>
Co-Authored-By: Ashwin Agate <ashwin.agate@suse.com>

Story: 2003023
Task: 23047

Change-Id: I8ed7a19c0565a0cd613866934eb2540ae46c847d