Upgrade Apache Kafka client
Currently in all Python Monasca components the copy of `kafka-python`
library in version 0.9.5 (released on Feb 16, 2016) is used. This
specification describes the process of upgrading the Apache Kafka client
to `confluent-kafka-python`. This will improve the performance and
reliability. Sticking with the old frozen client version is also
unacceptable in terms of security.

Change-Id: I59f3effcdba39199d61d70a201d8e760840d3627
Story: 2003705
Task: 26360